pyproject.toml
[project]
name = "pyjdbc2"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "phi", email = "phi.friday@gmail.com" }
]
dependencies = [
    "jpype1==1.2.1",
    "typing-extensions>=4.4.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ruff==0.5.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pyjdbc2"]
